Add custom hook for loginform
